Add files via upload
This is the second week internship project at motion cut Front-end Development. Completed the register form.
It asks for username, email, password and confirmation of password with validations.
This project device optimal that is works on all device because media query is used in the project .
The Register button confirms the  registration ones all the input fields are correctly filled.